feature: Add full mipmap generation support for packs with oddly sized textures
This specifically affects MCC as we have various textures that are not divisible by 16 (which is required for mipmap generation), this could be fixed by being more careful with asset creation but instead I figured I could just make clients not complain and just generate mipmaps regardless of the input size. I really just rewrote Mojang's code to be nullable so it works even if pixels are missing. Need to investigate if any performance mods override the functions and break the behaviour.